requirements.txt

# DEVOPS & DRIVERS
# ----------------

dramatiq[redis, watch]  # Task queue alternative to celery https://dramatiq.io
# coralogix-logger==2.0.2
hiredis==1.1.0  # https://github.com/redis/hiredis-py
# psycopg2==2.8.5 --no-binary psycopg2  # https://github.com/psycopg/psycopg2
redis==3.5.3  # https://github.com/andymccurdy/redis-py
# whitenoise==5.1.0  # https://github.com/evansd/whitenoise
django-dramatiq==0.9.1

# APPLICATION
# -----------
# amqp==2.6.0
# anyjson==0.3.3
# argon2-cffi==20.1.0
# asn1crypto==1.4.0
# certifi==2020.6.20
# cffi==1.14.1
channels==2.4.0
channels_redis[cryptography]==3.1.0
click==7.1.2
# chardet==3.0.4
# coreapi==2.3.3
# coreschema==0.0.4
# cryptography==3.0
django==3.1.2
# django-anymail==7.2
# django-appconf==1.0.4
# django-celery-beat==2.0.0
# django-compressor==2.4
# django-cors-headers==3.4.0
django-environ==0.4.5
# django-extra-fields==2.0.5
# django-filter==2.3.0
# django-guardian==2.3.0
# django-imagekit==4.0.2
# django-map-widgets==0.3.0
# django-modelcluster<6.0,>=5.0
# django-redis==4.12.1
# django-redisboard==4.1.1
# django-rest-auth==0.9.5
# django-storages[google]==1.9.1
# django-taggit<2.0,>=1.0
# django-timezone-field==4.0
# django-treebeard==4.3.1
# djangorestframework==3.11.0
# djangorestframework-gis==0.15.0
# djangorestframework-guardian==0.3.0
# djangorestframework-simplejwt==4.4.0
# draftjs-exporter==4.1.0
# drf-yasg==1.17.1
# html5lib==1.1
# idna==2.10
# lxml==4.5.2
# MarkupSafe==1.1.1
# model-mommy==2.0.0
# pilkit==2.0
# Pillow==7.1.2
# psycopg2-binary==2.7.7
# pycparser==2.20
# PyJWT==1.7.1
# python-crontab==2.5.1
# python-dateutil==2.8.1
# pytz==2020.1
# rcssmin==1.0.6
# requests==2.22.0
# ruamel.yaml==0.16.10
# six==1.15.0
# sqlparse==0.3.1
# Unidecode==1.1.1
# uritemplate==3.0.1
# urllib3==1.25.10
# vine==1.3.0
# wagtail==2.9.3
# wagtailfontawesome==1.2.1
# wagtailleafletwidget==1.0.1
# webencodings==0.5.1
# Willow<1.4,>=1.3
twined==0.0.12
octue==0.1.3
